server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pydantic==2.8.0
pyyaml==6.0.2
python-multipart==0.0.12
httpx==0.27.0
cryptography==43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
